Remove usages of elasticsearch.build plugin in non-production projects (#84961)
(cherry picked from commit dc3d3b5e1242693e429aa003942ecfa0404a5b30)
