+++ build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/test/RestTestBasePlugin.java
@@ -82,7 +82,7 @@ public class RestTestBasePlugin implements Plugin<Project> {
 
         project.getTasks().withType(StandaloneRestIntegTestTask.class).configureEach(t ->
         // if this a module or plugin, it may have an associated zip file with it's contents, add that to the test cluster
-        project.getPluginManager().withPlugin("elasticsearch.internal-es-plugin", plugin -> {
+        project.getPluginManager().withPlugin("elasticsearch.esplugin", plugin -> {
             TaskProvider<Zip> bundle = project.getTasks().withType(Zip.class).named("bundlePlugin");
             t.dependsOn(bundle);
             if (GradleUtils.isModuleProject(project.getPath())) {
