+++ qa/system-indices/build.gradle
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-apply plugin: 'elasticsearch.internal-es-plugin'
+apply plugin: 'elasticsearch.base-internal-es-plugin'
 apply plugin: 'elasticsearch.internal-java-rest-test'
 
 esplugin {
@@ -15,11 +15,6 @@ esplugin {
   classname 'org.elasticsearch.system.indices.SystemIndicesQA'
   licenseFile rootProject.file('licenses/SSPL-1.0+ELASTIC-LICENSE-2.0.txt')
   noticeFile rootProject.file('NOTICE.txt')
-}
-
-tasks.named("test").configure { enabled = false }
-tasks.named("javaRestTest").configure {
-  dependsOn "buildZip"
 }
 
 testClusters.configureEach {
