+++ test/external-modules/build.gradle
@@ -1,8 +1,8 @@
 
-import org.elasticsearch.gradle.internal.info.BuildParams;
+import org.elasticsearch.gradle.internal.info.BuildParams
 
 subprojects {
-  apply plugin: 'elasticsearch.internal-es-plugin'
+  apply plugin: 'elasticsearch.base-internal-es-plugin'
   apply plugin: 'elasticsearch.internal-yaml-rest-test'
 
   esplugin {
