+++ x-pack/plugin/ccr/qa/downgrade-to-basic-license/build.gradle
@@ -58,8 +58,3 @@ tasks.register("follow-cluster", RestIntegTestTask) {
 }
 
 tasks.named("check").configure { dependsOn "follow-cluster" }
-
-// We can't run in FIPS mode with a basic license
-tasks.withType(Test).configureEach {
-  onlyIf { BuildParams.inFipsJvm == false}
-}
