+++ x-pack/plugin/security/qa/operator-privileges-tests/build.gradle
@@ -1,4 +1,6 @@
-apply plugin: 'elasticsearch.internal-es-plugin'
+import org.elasticsearch.gradle.Version
+
+apply plugin: 'elasticsearch.base-internal-es-plugin'
 apply plugin: 'elasticsearch.internal-java-rest-test'
 
 esplugin {
