+++ x-pack/qa/runtime-fields/build.gradle
@@ -1,4 +1,4 @@
-apply plugin: 'elasticsearch.build'
+apply plugin: 'elasticsearch.java'
 
 dependencies {
   api project(":test:framework")
